requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.33.0
sqlalchemy==2.0.47
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic-settings==2.8.1

pytest==8.3.5
pytest-cov==5.0.0
ruff==0.15.4